platform-lite: review fixes — {message} on 400, pin CTE status, relocate test
- 400 prefix-reject body now carries a message key: mcp's assertSuccess
  parses non-2xx bodies as {message}, so the informative read-only text
  was collapsing to the generic 'Failed to fetch logs' fallback (error
  kept for shape consistency with the 200 SQL-error path)
- CTE-reject test pins status 200: the prefix gate's 400 message also
  matches /read-only/i, so unifying the rejection paths would otherwise
  leave the read-only transaction guard silently untested
- move src/management-api/debugging.test.ts -> test/clickhouse-logs.test.ts:
  platform-lite tests live under test/, and the src placement collided on
  basename with the existing test/debugging.test.ts
